User authentication

+++ controllers/expense.js
@@ -9,7 +9,7 @@ exports.postExpenseData = async (req, res, next) => {
         if (!amount || !description || !category) {
             res.status(500).json({ message: 'Missing some data' });
         }
-        let data = await Expense.create({ amount: amount, description: description, category: category });
+        let data = await req.user.createExpense({ amount: amount, description: description, category: category });
         res.status(201).json(data);
     } catch (error) {
         res.status(500).json({ message: error });
@@ -18,7 +18,7 @@ exports.postExpenseData = async (req, res, next) => {
 
 exports.getExpenseData = async (req, res, next) => {
     try {
-        let data = await Expense.findAll();
+        let data = await req.user.getExpenses();
         res.status(201).json(data);
     } catch (error) {
         res.status(500).json({ message: error });
@@ -28,7 +28,7 @@ exports.getExpenseData = async (req, res, next) => {
 exports.deleteExpenseData = async (req, res, next) => {
     try {
         const prodId = req.params.id;
-        await Expense.destroy({ where: { id: prodId } });
+        await Expense.destroy({ where: { id: prodId, userId: req.user.id } });
         res.status(201).json(prodId);
     } catch (error) {
         res.status(500).json({ message: error });
